+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fixed issue where custom layouts would have displays created for them automatically
 - Fixed issues where field layouts would not have their fields populated properly. [10237](https://github.com/craftcms/cms/issues/10237)
 - Fixed issues with fields not properly rebuilt
+- Fixed errors with missing fields
 
 ### Changed
 - Rendered displayers label